rearranged process function to be under functions, fixed functions and build menu accordingly.
