+++ threat/database/database_module.py
@@ -21,7 +21,7 @@ import sqlite3
 #   data = should be the data generated by the module. It will be unchanged when it is
 #       passed in and out of the database.
 
-def save_data(database, module, lvl1, lvl2, lvl3, data, host):
+def save_data(database, module, lvl1, lvl2, lvl3, host, data):
     connection = sqlite3.connect(database)
     cursor = connection.cursor()
     check_table = 'SELECT name FROM sqlite_master WHERE type=\'table\' AND name=\'{}\''.format(module)
