+++ threat/modules/recon/passive/dig.py
@@ -4,6 +4,7 @@ from database.database_module import save_data
 def dig(target):
     for host in target:
         host.lvl2='dig'
+        host.lvl3=''
         DIGSCAN = "dig "+host.name
         results_dig = subprocess.check_output(DIGSCAN, shell=True)
         data=results_dig.decode().replace("<<","").replace(">>","")
